LayoutTests/imported/w3c/web-platform-tests/dom/nodes/moveBefore/listed-form-element-reset-expected.txt



FAIL Listed form element form owner is reset when a new form is moved closer before the listed element assert_equals: button.form is correctly reset to form2 expected Element node <form id="form1"></form> but got Element node <form id="form1"></form>
